requirement.txt
gradio
transformers
torch
psycopg2-binary